feat(vfs): adjust the directory structure
