+++ modules/vfs/android/src/main/cpp/src/handler/uri.cc
@@ -20,7 +20,7 @@
  *
  */
 
-#include "handler/uri.h"
+#include "vfs/uri.h"
 
 #include "footstone/string_view_utils.h"
 #include "jni/jni_env.h"
